+++ client/components/DesktopContainer.js
@@ -1,0 +1,101 @@
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import Carousel from "nuka-carousel";
+import Footer from "./Footer";
+import HomepageHeading from "./HomePageHeading";
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+} from "semantic-ui-react";
+import { NavLink, withRouter } from "react-router-dom";
+
+/* Heads up!
+ * Neither Semantic UI nor Semantic UI React offer a responsive navbar, however, it can be implemented easily.
+ * It can be more complicated, but you can create really flexible markup.
+ */
+
+class DesktopContainer extends Component {
+  state = {};
+
+  hideFixedMenu = () => this.setState({ fixed: false });
+  showFixedMenu = () => this.setState({ fixed: true });
+
+  render() {
+    const { children } = this.props;
+    const { fixed } = this.state;
+
+    return (
+      <Responsive {...Responsive.onlyComputer}>
+        <Visibility
+          once={false}
+          onBottomPassed={this.showFixedMenu}
+          onBottomPassedReverse={this.hideFixedMenu}
+        >
+          <Segment
+            inverted
+            textAlign="center"
+            style={{ minHeight: 700, padding: "1em 0em" }}
+            vertical
+          >
+            <Menu
+              fixed={fixed ? "top" : null}
+              inverted={!fixed}
+              pointing={!fixed}
+              secondary={!fixed}
+              size="large"
+            >
+              <Container>
+                <Menu.Item as={NavLink} to="/">
+                  Home
+                </Menu.Item>
+                <Menu.Item exact as={NavLink} to="/howitworks">
+                  How It Works
+                </Menu.Item>
+                <Menu.Item as="a">Security</Menu.Item>
+                <Menu.Item as="a">Blog</Menu.Item>
+                <Menu.Item as="a">Support</Menu.Item>
+                <Menu.Item exact as={NavLink} to="/aboutus">
+                  About Us
+                </Menu.Item>
+                <Menu.Item position="right">
+                  <Button as="a" inverted={!fixed}>
+                    Log in
+                  </Button>
+                  <Button
+                    as={NavLink}
+                    to="/signup"
+                    inverted={!fixed}
+                    primary={fixed}
+                    style={{ marginLeft: "0.5em" }}
+                  >
+                    Sign Up
+                  </Button>
+                </Menu.Item>
+              </Container>
+            </Menu>
+            <HomepageHeading />
+          </Segment>
+        </Visibility>
+
+        {children}
+      </Responsive>
+    );
+  }
+}
+
+DesktopContainer.propTypes = {
+  children: PropTypes.node,
+};
+
+export default withRouter(DesktopContainer);
