+++ client/components/HomePage.js
@@ -15,221 +15,11 @@ import {
   Responsive,
   Segment,
   Sidebar,
-  Visibility
+  Visibility,
 } from "semantic-ui-react";
 import { NavLink, withRouter } from "react-router-dom";
-
-/* eslint-disable react/no-multi-comp */
-/* Heads up! HomepageHeading uses inline styling, however it's not the best practice. Use CSS or styled components for
- * such things.
- */
-
-const HomepageHeading = ({ mobile }) => (
-  <div text>
-    <section id="carousel">
-      <Header
-        className="carousel-text"
-        as="h1"
-        content="Imagine-a-Company"
-        inverted
-        style={{
-          fontSize: mobile ? "2em" : "4em",
-          fontWeight: "normal",
-          marginBottom: 0,
-          marginTop: mobile ? "1.5em" : "3em"
-        }}
-      />
-
-      <Carousel>
-        <img
-          className="carousel-image"
-          src="../../../../../../pictures/gold.jpg"
-        />
-        <img
-          className="carousel-image"
-          src="../../../../../../pictures/waterplants.jpg"
-        />
-      </Carousel>
-      <Header
-        className="carousel-text"
-        as="h2"
-        content="Do whatever you want when you want to."
-        inverted
-        style={{
-          fontSize: mobile ? "1.8em" : "2.0em",
-          fontWeight: "normal",
-          marginTop: mobile ? "0.5em" : "1.5em"
-        }}
-      />
-    </section>
-    <Button primary size="huge">
-      Get Started
-      <Icon name="right arrow" />
-    </Button>
-  </div>
-);
-
-HomepageHeading.propTypes = {
-  mobile: PropTypes.bool
-};
-
-/* Heads up!
- * Neither Semantic UI nor Semantic UI React offer a responsive navbar, however, it can be implemented easily.
- * It can be more complicated, but you can create really flexible markup.
- */
-class DesktopContainer extends Component {
-  state = {};
-
-  hideFixedMenu = () => this.setState({ fixed: false });
-  showFixedMenu = () => this.setState({ fixed: true });
-
-  render() {
-    const { children } = this.props;
-    const { fixed } = this.state;
-
-    return (
-      <Responsive {...Responsive.onlyComputer}>
-        <Visibility
-          once={false}
-          onBottomPassed={this.showFixedMenu}
-          onBottomPassedReverse={this.hideFixedMenu}
-        >
-          <Segment
-            inverted
-            textAlign="center"
-            style={{ minHeight: 700, padding: "1em 0em" }}
-            vertical
-          >
-            <Menu
-              fixed={fixed ? "top" : null}
-              inverted={!fixed}
-              pointing={!fixed}
-              secondary={!fixed}
-              size="large"
-            >
-              <Container>
-                <Menu.Item as="a" active>
-                  Home
-                </Menu.Item>
-                <Menu.Item as="a">How It Works</Menu.Item>
-                <Menu.Item as="a">Security</Menu.Item>
-                <Menu.Item as="a">Blog</Menu.Item>
-                <Menu.Item as="a">Support</Menu.Item>
-                <Menu.Item as="a">About Us</Menu.Item>
-                <Menu.Item position="right">
-                  <Button as="a" inverted={!fixed}>
-                    Log in
-                  </Button>
-                  <Button
-                    as={NavLink}
-                    to="/signup"
-                    inverted={!fixed}
-                    primary={fixed}
-                    style={{ marginLeft: "0.5em" }}
-                  >
-                    Sign Up
-                  </Button>
-                </Menu.Item>
-              </Container>
-            </Menu>
-            <HomepageHeading />
-          </Segment>
-        </Visibility>
-
-        {children}
-      </Responsive>
-    );
-  }
-}
-
-DesktopContainer.propTypes = {
-  children: PropTypes.node
-};
-
-class MobileContainer extends Component {
-  state = {};
-
-  handlePusherClick = () => {
-    const { sidebarOpened } = this.state;
-
-    if (sidebarOpened) this.setState({ sidebarOpened: false });
-  };
-
-  handleToggle = () =>
-    this.setState({ sidebarOpened: !this.state.sidebarOpened });
-
-  render() {
-    const { children } = this.props;
-    const { sidebarOpened } = this.state;
-
-    return (
-      <Responsive {...Responsive.onlyMobile}>
-        <Sidebar.Pushable>
-          <Sidebar
-            as={Menu}
-            animation="uncover"
-            inverted
-            vertical
-            visible={sidebarOpened}
-          >
-            <Menu.Item as="a" active>
-              Home
-            </Menu.Item>
-            <Menu.Item as="a">How It Works</Menu.Item>
-            <Menu.Item as="a">Security</Menu.Item>
-            <Menu.Item as="a">Blog</Menu.Item>
-            <Menu.Item as="a">Support</Menu.Item>
-            <Menu.Item as="a">Log in</Menu.Item>
-            <Menu.Item as={NavLink} to="/signup">
-              Sign Up
-            </Menu.Item>
-          </Sidebar>
-
-          <Sidebar.Pusher
-            dimmed={sidebarOpened}
-            onClick={this.handlePusherClick}
-            style={{ minHeight: "100vh" }}
-          >
-            <Segment
-              inverted
-              textAlign="center"
-              style={{ minHeight: 350, padding: "1em 0em" }}
-              vertical
-            >
-              <Container>
-                <Menu inverted pointing secondary size="large">
-                  <Menu.Item onClick={this.handleToggle}>
-                    <Icon name="sidebar" />
-                  </Menu.Item>
-                  <Menu.Item position="right">
-                    <Button as="a" inverted>
-                      Log in
-                    </Button>
-                    <Button
-                      as={NavLink}
-                      to="/signup"
-                      inverted
-                      style={{ marginLeft: "0.5em" }}
-                    >
-                      Sign Up
-                    </Button>
-                  </Menu.Item>
-                </Menu>
-              </Container>
-              <HomepageHeading mobile />
-            </Segment>
-
-            {children}
-          </Sidebar.Pusher>
-        </Sidebar.Pushable>
-      </Responsive>
-    );
-  }
-}
-
-MobileContainer.propTypes = {
-  children: PropTypes.node
-};
+import MobileContainer from "./MobileContainer";
+import DesktopContainer from "./DesktopContainer";
 
 const ResponsiveContainer = ({ children }) => (
   <div>
@@ -239,7 +29,7 @@ const ResponsiveContainer = ({ children }) => (
 );
 
 ResponsiveContainer.propTypes = {
-  children: PropTypes.node
+  children: PropTypes.node,
 };
 
 const HomePage = () => (
