+++ client/components/MobileContainer.js
@@ -1,0 +1,108 @@
+import PropTypes from "prop-types";
+import React, { Component } from "react";
+import Carousel from "nuka-carousel";
+import Footer from "./Footer";
+import HomepageHeading from "./HomePageHeading";
+import {
+  Button,
+  Container,
+  Divider,
+  Grid,
+  Header,
+  Icon,
+  Image,
+  List,
+  Menu,
+  Responsive,
+  Segment,
+  Sidebar,
+  Visibility,
+} from "semantic-ui-react";
+import { NavLink, withRouter } from "react-router-dom";
+
+class MobileContainer extends Component {
+  state = {};
+
+  handlePusherClick = () => {
+    const { sidebarOpened } = this.state;
+
+    if (sidebarOpened) this.setState({ sidebarOpened: false });
+  };
+
+  handleToggle = () =>
+    this.setState({ sidebarOpened: !this.state.sidebarOpened });
+
+  render() {
+    const { children } = this.props;
+    const { sidebarOpened } = this.state;
+
+    return (
+      <Responsive {...Responsive.onlyMobile}>
+        <Sidebar.Pushable>
+          <Sidebar
+            as={Menu}
+            animation="uncover"
+            inverted
+            vertical
+            visible={sidebarOpened}
+          >
+            <Menu.Item as={NavLink} to="/">
+              Home
+            </Menu.Item>
+            <Menu.Item as="a">How It Works</Menu.Item>
+            <Menu.Item as="a">Security</Menu.Item>
+            <Menu.Item as="a">Blog</Menu.Item>
+            <Menu.Item as="a">Support</Menu.Item>
+            <Menu.Item as="a">Log in</Menu.Item>
+            <Menu.Item as={NavLink} to="/signup">
+              Sign Up
+            </Menu.Item>
+          </Sidebar>
+
+          <Sidebar.Pusher
+            dimmed={sidebarOpened}
+            onClick={this.handlePusherClick}
+            style={{ minHeight: "100vh" }}
+          >
+            <Segment
+              inverted
+              textAlign="center"
+              style={{ minHeight: 350, padding: "1em 0em" }}
+              vertical
+            >
+              <Container>
+                <Menu inverted pointing secondary size="large">
+                  <Menu.Item onClick={this.handleToggle}>
+                    <Icon name="sidebar" />
+                  </Menu.Item>
+                  <Menu.Item position="right">
+                    <Button as="a" inverted>
+                      Log in
+                    </Button>
+                    <Button
+                      as={NavLink}
+                      to="/signup"
+                      inverted
+                      style={{ marginLeft: "0.5em" }}
+                    >
+                      Sign Up
+                    </Button>
+                  </Menu.Item>
+                </Menu>
+              </Container>
+              <HomepageHeading mobile />
+            </Segment>
+
+            {children}
+          </Sidebar.Pusher>
+        </Sidebar.Pushable>
+      </Responsive>
+    );
+  }
+}
+
+MobileContainer.propTypes = {
+  children: PropTypes.node,
+};
+
+export default withRouter(MobileContainer);
