+++ client/components/index.js
@@ -8,3 +8,5 @@ export { default as UserHome } from "./user-home";
 export { Login, Signup } from "./auth-form";
 export { default as HomePage } from "./HomePage";
 export { default as SignUp } from "./SignUp";
+export { default as AboutUs } from "./AboutUs";
+export { default as HowItWorks } from "./HowItWorks";
