+++ public/style.css
@@ -28,7 +28,7 @@ form div {
 #carousel .carousel-text {
   position: absolute;
   top: 45vh;
-  opacity: .7;
+  opacity: 0.7;
   color: rgb(17, 23, 34);
   text-shadow: 2px 2px black;
   z-index: 1;
@@ -44,4 +44,10 @@ form div {
   /* transition: opacity 1.5s; */
 }
 
+#aboutUsBackground {
+  background: linear-gradient(#fdfcfb, #e2d1c3);
+}
 
+.ui.cards {
+  flex-wrap: initial;
+}
